Fix new book not being added due to incorrect handling of ISBN check
Previous implementation of `saveBook` in `BookService` made use of `findByISBN` to check for ISBN conflicts when adding a new book. Due to `findByISBN` natural behavior, the method threw an exception whenever the ISBN was not found in the list, which is always the case for new books. This commit implements `existsByISBN` in `BookService`, which fixes that.