Fix search within the 'genre' field
Previously, if a book belonged to a genre with two words, like 'science fiction', the search would only return results if the full genre name was given. This change splits the genre into individual words and indexes the book by each word in this field, increasing the chances of finding matching books.